requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.32.0
azure-storage-blob==12.23.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docarray==0.40.0
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
fonttools==4.54.1
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
isodate==0.7.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-experimental==0.3.3
langchain-text-splitters==0.3.2
langsmith==0.1.147
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.9.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.57.0
openpyxl==3.1.5
orjson==3.10.12
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.7
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.28.2
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
streamlit==1.39.0
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.18.3
yfinance==0.2.49
zipp==3.20.2
